Import the V4 robots, and make the import pipeline run outside its author's tree
Ports v4_s1_v4_ll's robots. The RobotDefinition YAMLs and the import scripts
already existed from the collaborator's commit; every input they needed was
absent, all of it deliberately gitignored (USDs, six URDF checkouts, ROS).

Four changes make the pipeline run here:

* A ~30 line ament_index_python shim. pip's xacro resolves $(find pkg) through
  that module, which ships only with ROS 2 and is not on PyPI. The shim reads
  the exact ament index layout `register_package` already builds, so requiring
  a full ROS Humble install for one lookup is no longer the price of entry.
* XACRO_BIN / XACRO_PYTHON overrides, replacing hardcoded /opt/ros/humble paths.
* Per-robot source detection: a missing checkout skips that robot instead of
  failing the run. That is what lets us leave out FANUC's ~614 MB mesh source,
  which only the H* tasks' checkpoint arms use.
* resolved_urdf/ is created explicitly. It had been relying on the first xacro()
  call as a side effect, and that call is now skippable.

Imported, registered, and loaded: v4_ridgeback_ur5 (30 links, 11 joints, eef
ur_arm_tool0), v4_jackal, v4_crazyflie_cf2x.

Loading the Ridgeback is what found the rest. OmniGibson's manipulation path
assumes a finger gripper in four separate places, and a suction end effector
trips every one:

  finger_link_names        falls back to {} -> KeyError('0') at init
  is_grasping              "physical" mode reads gripper_<arm>
  gripper_control_idx      th.tensor([]) is float32 -> cannot index with it
  _handle_assisted_grasping runs every physics step, reads gripper_<arm>

Three are answered by configuration and are now in
coop2/robot_configs/v4_ridgeback_ur5_primitives.yaml: no fingers declared
explicitly, grasping_mode sticky, disable_grasp_handling true -- the last being
the documented switch for a caller that drives grasps itself, which the symbolic
layer does through _establish_grasp and release_grasp_immediately.

The fourth is a real bug and is fixed in robot.py: joint index tensors are built
from a list comprehension, and an empty list yields float32. An empty list is
legitimate for a robot with no fingers, so that one is now dtype=th.int.
Non-empty cases are unchanged. Four sibling properties have the same latent
issue; only the one that actually fires is touched.

None of this needs the symbolic grasp to change. It teleports the object to the
end effector and welds a FixedJoint, which is what a suction cup is -- the
finger-contact path it never uses is the only thing that was in the way.

+++ OmniGibson/omnigibson/robots/robot.py
@@ -2829,7 +2829,15 @@ class Robot(USDObject, GymObservable):
         """
         assert self.is_manipulation
         return {
-            arm: th.tensor([list(self.joints.keys()).index(name) for name in self.finger_joint_names[arm]])
+            # dtype is explicit because these are used as indices and the list
+            # is legitimately empty for a robot with no fingers -- a suction
+            # gripper, say. th.tensor([]) is float32, and indexing a tensor with
+            # it raises "tensors used as indices must be long, int, byte or
+            # bool". Non-empty cases are unaffected: joint indices are ints.
+            arm: th.tensor(
+                [list(self.joints.keys()).index(name) for name in self.finger_joint_names[arm]],
+                dtype=th.int,
+            )
             for arm in self.arm_names
         }
